+++ src/index.css
@@ -11,6 +11,7 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  scroll-behavior: smooth;
   /* background: var(--bgColor); */
 }
 body {
